+++ src/Endpoints/RetailerAPI/Products.php
@@ -5,14 +5,16 @@ namespace Budgetlens\BolRetailerApi\Endpoints\RetailerAPI;
 use Budgetlens\BolRetailerApi\Endpoints\BaseEndpoint;
 use Budgetlens\BolRetailerApi\Requests\ListProductsRequest;
 use Budgetlens\BolRetailerApi\Resources\FiltersList;
+use Budgetlens\BolRetailerApi\Resources\Product\Asset;
 use Budgetlens\BolRetailerApi\Resources\ProductList;
-use Budgetlens\BolRetailerApi\Resources\Retailer;
+use Budgetlens\BolRetailerApi\Resources\Subscription;
 use Budgetlens\BolRetailerApi\Support\Str;
+use Illuminate\Support\Collection;
 
 class Products extends BaseEndpoint
 {
 
-    public function list(ListProductsRequest $request)
+    public function list(ListProductsRequest $request): null | ProductList
     {
         $response = $this->performApiCall(
             'POST',
@@ -30,9 +32,8 @@ class Products extends BaseEndpoint
         return null;
     }
 
-    public function listFilters(ListProductsRequest $request)
+    public function listFilters(ListProductsRequest $request): null | FiltersList
     {
-
         $parameters = collect($request->toArray())
             ->map(function ($data, $key) {
                 return [Str::snake($key, '-') => $data];
@@ -56,13 +57,26 @@ class Products extends BaseEndpoint
         return null;
     }
 
-    public function get(
-        string | int $retailerId,
-    ): Retailer {
+    public function getAssets(string $eancode, null | string $usage = null): Collection
+    {
+        $parameters = collect([
+            'usage' => $usage
+        ])
+            ->reject(function ($value) {
+                return is_null($value);
+            })->all();
+
         $response = $this->performApiCall(
             'GET',
-            "retailers/{$retailerId}"
+            "products/{$eancode}/assets" . $this->buildQueryString($parameters)
         );
-        return new Retailer(collect($response));
+
+        $collection = new Collection();
+
+        collect($response->assets)->each(function ($item) use ($collection) {
+            $collection->push(new Asset($item));
+        });
+
+        return $collection;
     }
 }